+++ src/background/state/domain.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+// ============== THEMES ==============
+
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 #[serde(default, rename_all = "camelCase")]
 pub struct ThemeCss {
@@ -23,4 +25,24 @@ pub struct ThemeInfo {
 pub struct Theme {
     pub info: ThemeInfo,
     pub styles: ThemeCss,
+}
+
+// ============== PLACEHOLDERS ==============
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+#[serde(default, rename_all = "camelCase")]
+pub struct PlaceholderInfo {
+    pub display_name: String,
+    pub author: String,
+    pub description: String,
+    pub filename: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+#[serde(default, rename_all = "camelCase")]
+pub struct Placeholder {
+    pub info: PlaceholderInfo,
+    pub left: Vec<serde_yaml::Value>,
+    pub center: Vec<serde_yaml::Value>,
+    pub right: Vec<serde_yaml::Value>,
 }