+++ src/background/state/infrastructure.rs
@@ -1,12 +1,26 @@
 use itertools::Itertools;
 
-use super::{domain::Theme, themes::THEME_MANAGER};
+use super::{
+    domain::{Placeholder, Theme},
+    placeholders::PLACEHOLDERS_MANAGER,
+    themes::THEME_MANAGER,
+};
 
 #[tauri::command]
 pub fn state_get_themes() -> Vec<Theme> {
     THEME_MANAGER
         .lock()
         .themes()
+        .values()
+        .cloned()
+        .collect_vec()
+}
+
+#[tauri::command]
+pub fn state_get_placeholders() -> Vec<Placeholder> {
+    PLACEHOLDERS_MANAGER
+        .lock()
+        .placeholders()
         .values()
         .cloned()
         .collect_vec()